% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{bioDev}
\alias{bioDev}
\title{Initial relative biomass deviations}
\usage{
bioDev(TimeAreaObj, StochasticObj = NULL)
}
\arguments{
\item{TimeAreaObj}{A TimeArea object}

\item{StochasticObj}{A Stochastic object}
}
\description{
Initial relative biomass deviations
}
