% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{cpueDev}
\alias{cpueDev}
\title{Initial cpue deviations - special function for projectionStrategy}
\usage{
cpueDev(TimeAreaObj, StrategyObj)
}
\arguments{
\item{TimeAreaObj}{A TimeArea object}

\item{StrategyObj}{A Stochastic object}
}
\description{
Initial cpue deviations - special function for projectionStrategy
}
