% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{effortImpErrorDev}
\alias{effortImpErrorDev}
\title{Effort implementation error in projection - special function for projectionStrategy}
\usage{
effortImpErrorDev(TimeAreaObj, StrategyObj)
}
\arguments{
\item{TimeAreaObj}{A TimeArea object}

\item{StrategyObj}{A Stochastic object}
}
\description{
Effort implementation error in projection - special function for projectionStrategy
}
