% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEWrappers.R
\name{evalMSE}
\alias{evalMSE}
\title{Population dynamics wrapper called by runProjection}
\usage{
evalMSE(inputObject)
}
\arguments{
\item{inputObject}{A list of objects passed from runProjection}
}
\description{
Contains population dynamics equations. Should not be run directly, instead called by runProjection
}
