% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTGForFishKit.R
\docType{package}
\name{fishSimGTGForFishKit}
\alias{fishSimGTGForFishKit-package}
\alias{fishSimGTGForFishKit}
\title{Numerical simulations of fish population dynamics}
\description{
Conducts projections, MSE, and equilibrium conditions for a specified fish life history.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natureanalytics-ca/fishSimGTGForFishKit}
  \item Report bugs at \url{https://github.com/natureanalytics-ca/fishSimGTGForFishKit/issues}
}

}
\author{
\strong{Maintainer}: William Harford \email{bill@natureanalytics.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
