% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managementStrategies.R
\name{fixedStrategy}
\alias{fixedStrategy}
\title{Historical fishing pressure}
\usage{
fixedStrategy(phase, dataObject)
}
\arguments{
\item{phase}{Management procedures are coded in three phases: 1 - data collection, 2 - a decision making process, 3 - conversion of that process into annual F}

\item{dataObject}{The needed inputs to the management procedure}
}
\description{
Historical fishing pressure
}
