% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTGForFishKit.R
\docType{data}
\name{gtgSimExample}
\alias{gtgSimExample}
\title{GTG sim example}
\format{
An S4 object of class LBSPRarray
}
\source{
Simulated data
}
\usage{
gtgSimExample
}
\description{
The data object LifeHistoryExample is used in gtgYPRWrapper to produce YPR & SPR arrays.
}
\keyword{datasets}
