% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesS4.R
\docType{class}
\name{LBSPRarray-class}
\alias{LBSPRarray-class}
\title{LBSPR YPR object}
\description{
An S4 object that holds the output of YPR analysis in a standardized format.
#' @importFrom methods new
}
