% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTGForFishKit.R
\docType{data}
\name{lbsprSimExample}
\alias{lbsprSimExample}
\title{LBSPR sim example}
\format{
An S4 object of class LBSPRarray
}
\source{
Simulated data
}
\usage{
lbsprSimExample
}
\description{
The data object LifeHistoryExample is used in lsbsprSimWrapper to produce YPR & SPR arrays.
}
\keyword{datasets}
