% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YPR.R
\name{lbsprSimWrapperAbsel}
\alias{lbsprSimWrapperAbsel}
\title{LBSPR using 'absel' methodeumetric simulation}
\usage{
lbsprSimWrapperAbsel(
  LifeHistoryObj,
  binWidth = 1,
  binMin = 0,
  LcStep = 1,
  F_MStep = 0.2,
  waitName = NULL,
  hostName = NULL
)
}
\arguments{
\item{LifeHistoryObj}{A life history object.}

\item{binWidth}{LBSPR length bin width, default value is 1.}

\item{binMin}{LBSPR binMin, default is 0}

\item{LcStep}{Length step size in cm for sequence of length at vulnerability. Approx. knife edge vul. SL50 = Lc, SL95 = Lc + 1}

\item{F_MStep}{F/M ratio step size for sequence of F_M}

\item{waitName}{When used within a shiny app, this function can update a host from the waiter package. See example.}

\item{hostName}{When used within a shiny app, this function can update a host from the waiter package. See example.}
}
\description{
For diagnostic purposes only. See lbsprSimWrapper for details.
}
