% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LH_to_dataframe}
\alias{LH_to_dataframe}
\title{Life History object to data frame}
\usage{
LH_to_dataframe(LifeHistoryObj, digits = 3)
}
\arguments{
\item{LifeHistoryObj}{A life history object.}

\item{digits}{Integer indicating the number of decimal places (round). Applied as significant digits (signif) to L-W alpha.}
}
\description{
Converts S4 life history object to data frame with nice variable names
}
\examples{
library(tidyverse)
LH_to_dataframe(LifeHistoryExample)
}
