% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{LHwrapper}
\alias{LHwrapper}
\title{Life history wrapper for sub-cohorts}
\usage{
LHwrapper(
  LifeHistoryObj,
  TimeAreaObj,
  stepsPerYear = 1,
  doPlot = FALSE,
  wd = NULL,
  imageName = NULL,
  dpi = 300
)
}
\arguments{
\item{LifeHistoryObj}{A life history object.}

\item{TimeAreaObj}{A time-area object}

\item{stepsPerYear}{Creates multiple time steps per year allowing finer scale length composition.}

\item{doPlot}{Creates a basic plot to visualize outcomes. Useful for ensuring parameter selections are sensible.}

\item{wd}{A working directly where the output of runProjection is saved}

\item{imageName}{Character. A name for the resulting plot(s)}

\item{dpi}{Resolution in dots per inch of the resulting saved chart.}
}
\description{
Creates the necessary age-based vectors describing life history of sub-cohorts
}
\examples{
ta<-new("TimeArea")
ta@gtg<-13
LHwrapper(LifeHistoryObj = LifeHistoryExample, TimeAreaObj=ta)
}
