% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{lifehistoryDev}
\alias{lifehistoryDev}
\title{Set of life history params}
\usage{
lifehistoryDev(TimeAreaObj, StochasticObj)
}
\arguments{
\item{TimeAreaObj}{A TimeArea object}

\item{StochasticObj}{A Stochastic object}
}
\description{
Set of life history params
}
