% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTGForFishKit.R
\docType{data}
\name{LifeHistoryExample}
\alias{LifeHistoryExample}
\title{Life history example}
\format{
An S4 object of class LifeHistory
}
\source{
Simulated data
}
\usage{
LifeHistoryExample
}
\description{
A LifeHistory object, prepopulated with an example fish life history.
}
\details{
This life history is based on the Beverton-Holt life history invariants of M/K=1.5 and Lm/Linf=0.66
}
\keyword{datasets}
