% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{MoveMat}
\alias{MoveMat}
\title{Move matrix calculations for cohort}
\usage{
MoveMat(Surv_in, Move_in, area_in)
}
\arguments{
\item{Surv_in}{The output of the SurvMat function}

\item{Move_in}{A matrix of migration rates of dimensions areas x areas. Obtained from TimeArea object.}

\item{area_in}{The area under evaluation (e.g., 1)}
}
\description{
Move matrix calculations for cohort
}
