% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managementStrategies.R
\name{projectionStrategy}
\alias{projectionStrategy}
\title{Static projections of combinations of effort changes, bag limit, and/or spatial closures. Also used in combination with ProFisheryObj (e.g. size limit change) to project temporal dynamics of size limits.}
\usage{
projectionStrategy(phase, dataObject)
}
\arguments{
\item{phase}{Management procedures are coded in three phases: 1 - data collection, 2 - a decision making process, 3 - conversion of that process into annual F}

\item{dataObject}{The needed inputs to the management procedure}
}
\description{
The Strategy object should be specified as follows. (1) Set Strategy@projectionYears to the number of forward projection years you wish to simulate. (2) Strategy@projectionName = "projectionStrategy".
(3)  Strategy@projectionParams should be a list with two items. First items is a vector of length areas containing bag limit. For no bag limit use -99.
The bag limit should be thought of as take per unit time (e.g. day) and basically acts like a CPUE threshold.
The effect of bag limit is calculated against the historical CPUE (e.g. in same units of take per unit time) in the Stochastic object historicalCPUE. Make sure that the bag limit and historicalCPUE are consistent with historicalCPUEType (e.g., biomass or abundance (numbers)) and this parameter is used in determing the effect of fish biomass or abundance on CPUE.
The second item in the list is a matrix of nrows = projectionYears and ncols = areas that contains value multipiers of initial equilibrium fishing effort. This allows projection of effort reduction and of marine reserves via setting effort to 0.
}
