% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEWrappers.R
\name{readProjection}
\alias{readProjection}
\title{Read in data from existing projection or MSE model}
\usage{
readProjection(wd, fileName)
}
\arguments{
\item{wd}{A working directly where output is saved. Required}

\item{fileName}{A file name previously used to create output. Required}
}
\description{
Function for running projections or MSE
}
