% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{recDev}
\alias{recDev}
\title{Recuitment deviations}
\usage{
recDev(LifeHistoryObj, TimeAreaObj, StochasticObj, StrategyObj = NULL)
}
\arguments{
\item{LifeHistoryObj}{A life history object.}

\item{TimeAreaObj}{A TimeArea object}

\item{StrategyObj}{A Strategy object}
}
\description{
Recuitment deviations
}
