% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{recruit}
\alias{recruit}
\title{Stock-recuitment function}
\usage{
recruit(LifeHistoryObj, B0 = 0, stock = 0, forceR = FALSE, Rforced = 0)
}
\arguments{
\item{LifeHistoryObj}{A life history object.}

\item{B0}{Unfished spawning stock biomass}

\item{stock}{the current spawning stock biomass}

\item{forceR}{A logical that overides the S-R function and simply returns a specified number of recruits. Default is FALSE}

\item{Rforced}{A fixed number of recruits to return. Only used when forceR is TRUE}
}
\description{
Stock-recuitment function
}
