% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{selDev}
\alias{selDev}
\title{Set of selectivity params}
\usage{
selDev(TimeAreaObj, HistFisheryObj, ProFisheryObj_list = NULL, StochasticObj)
}
\arguments{
\item{TimeAreaObj}{A TimeArea object}

\item{HistFisheryObj}{A fishery object for historical period}

\item{ProFisheryObj_list}{A fishery object list for projection period}

\item{StochasticObj}{A Stochastic object}
}
\description{
Set of selectivity params
}
