% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{solveD}
\alias{solveD}
\title{Equilibrium conditions for sub-cohorts}
\usage{
solveD(
  lh,
  sel,
  doFit = FALSE,
  F_in = NULL,
  D_type = NULL,
  D_in = NULL,
  doPlot = FALSE
)
}
\arguments{
\item{lh}{An object produced by LHWrapper.}

\item{sel}{An object produced by selWrapper}

\item{doFit}{Logical. When TRUE, estimates equilibrium fishing mortality based on input D_in. Ignores F_in. Default is FALSE}

\item{F_in}{Equilibrium fishing mortality rate. Used to calculate equilibrium conditions of the stock. Ignored when doFit = TRUE}

\item{D_type}{When doFit = TRUE, specifies type of equilibrium state metric that is specified in D_in (e.g., SSB depletion or SPR).}

\item{D_in}{When doFit = TRUE, specifies value of equilibrium state. Must be SSB depletion or SPR both with value between 0 and 1}

\item{doPlot}{Equilibrium length composition}
}
\description{
Creates the necessary age-based vectors equilibrium abundance, biomass and catch for sub-cohorts
}
