% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesS4.R
\docType{class}
\name{Strategy-class}
\alias{Strategy-class}
\title{Strategy object}
\arguments{
\item{title}{A title for the object, useful for describing the strategy}

\item{projectionYears}{Number of forward projection years to simulate}

\item{projectionName}{String. The name of projection method to apply. This is the name of the projection function}

\item{projectionParams}{List. List structure follows specification of the projection function specified in projectionName}
}
\description{
An S4 object that holds descriptions of projections to be made, including a harvest strategy
}
\details{
Details of projection to be made
}
