% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{SurvMat}
\alias{SurvMat}
\title{Surival matrix calculations for cohort}
\usage{
SurvMat(ageClasses, M_in, F_in, S_in)
}
\arguments{
\item{ageClasses}{Number of age classes. Typically max age + 1, since recruitment occurs at age 0}

\item{M_in}{Natural mortality rate}

\item{F_in}{Fishing mortality rate}

\item{S_in}{Selectivity, specifically the removal vector from selWrapper}
}
\description{
Surival matrix calculations for cohort
}
