% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesS4.R
\docType{class}
\name{TimeArea-class}
\alias{TimeArea-class}
\title{Time-area object}
\arguments{
\item{title}{A title for the object, useful for displaying the contents of the object}

\item{gtg}{Number of growth-type groups}

\item{areas}{Number of areas in the model, must be greater than 1.}

\item{recArea}{A vector of length areas. Fraction of recruitment to each area with values summing to 1.}

\item{move}{A matrix of migration rates of dimensions areas x areas}

\item{iterations}{Number of iterations to run}

\item{historicalYears}{Number of years to simulate historical dynamics}

\item{historicalBio}{Number greater than 0 and less than 1. Model assumes we are dealing with an already exploited fish population}

\item{historicalBioType}{String. The type of historical biomass state, options are: 'relB' or 'SPR'.}

\item{historicalEffort}{A matrix of nrows = historicalYears and ncols = areas that contains value multipiers of initial equilibrium fishing effort}
}
\description{
An S4 object that holds descriptions of time step, gtg, and area params
}
\details{
Inputs for number of gtg, time step, and areas
}
